date added animation added looks ugly

+++ app/src/main/java/ru/angelovich/chartapplication/data/ChartData.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class ChartData {
     public ArrayList<ChartLine> lines;
-    public int[] xAxis;
+    public long[] xAxis;
     public int size;
 }
